Convert check_deploy_timeout to use process_event
Convert check_deploy_timeout to use the async callback within
process_event, rather than setting it up "manually".

This change also modifies check_deploy_timeout to use an error
handler with the async callback in case a worker is unavailable for
the task.

This is the first instance of using process_event to transition to a
-FAIL status.

Change-Id: I1788e96f084b3c791a45ab710aeb63e6b9ab08a5
Related-to: blueprint new-ironic-state-machine