Ensure that image link points to the correct image
If an image is cached for a node, and the image changes, Ironic won't
notice because we only check if the link exists. In this change it's
proposed to check in the image_cache if the image is up to date.

Closes-bug: #1405448
Co-Authored-By: Vladyslav Drok <vdrok@mirantis.com>
(cherry picked from commit 2f0c8cd8d8d8b2846b641b518b742ca051425028)

Conflicts:
	ironic/drivers/modules/image_cache.py
	ironic/tests/drivers/test_image_cache.py

Change-Id: If67c7e60eb6d048fa3aa03469d7904c6c233ff89