Fix policy enforcement to properly detect admin
We were previously checking policy during the context creation hook
using a dict that the policy engine understood.  This was changed to
passing in a dict of raw headers, which the engine cannot use when
enforcing the new 'admin_api' rule, which depends on a dict of
sanitized keys.  This reverts to creating a similar dict as before that
is more complete.  Also, adds some needed test coverage around this stuff.

Change-Id: Ia98591bc388e8600c1955e2f9953eab654510305
Closes-bug: #1408811