Fix RPCService and Ironic Conductor so they shut down gracefully
Ironic Conductor Manager now waits on its greenpool workers to finish
before del_host completes. This will ensure more-correct behavior, as
well as giving workers the opportunity to release locks before the
RPCService stops. Added a manager test to confirm del_host waits on the
manager's greenpool.

This modifies RPCService in two important ways:

1) The order of shutdown operations in RPCService.stop is changed to be
more correct. First, the server is stopped to prevent new requests from
arriving. Second, the manager is instructed to inform the cluster that
this conductor is shutting down. Finally, the conductor service stop()
method is called with graceful set to True to allow any tasks the
service currently has running to finish normally.

This is in contrast to the previous order of operations where the
service was told to immediately stop before stopping the server. This
could allow requests to arrive while the service has already stopped,
which is not ideal. Additionally, Service.stop was called in a way
which caused existing/running service tasks to immediately exit, which
could result in bad node or database state, necessitating manual
operator intervention.

2) The underlying Service object is told to shutdown gracefully, which
will allow tasks to complete before RPCService completely shuts down.

Change-Id: I639f2fa7b9349822fca9180966e0932224fa70a2
Closes-Bug: #1382698