Agent driver: no-op heartbeat for maintenanced node
There may be a race where a node begins deploy or cleaning, but is
maintenanced (for whatever reason) before the ramdisk boots and reports
back. A node in maintenance mode should not be touched by Ironic.

No-op the heartbeat method when a node is in maintenance mode. Rather
than failing immediately, allow the agent to continue heartbeating with
the hope that it is taken out of maintenance mode soon.

Co-Author: Jim Rollenhagen <jim@jimrollenhagen.com>
Closes-Bug: #1436569
Change-Id: Ib461e1c2652a4a55c112b62eea9b3d5492c33b63