Add MANAGEABLE state and associated transitions
This patch adds the MANAGEABLE state, and associated transitions to and
from AVAILABLE, as well as a new RPC method to initiate transitions
between provisioning states.

It also adds a new VERBS entity to ironic/common/states.py to hold the
action-phrases, as described in blueprint new-ironic-state-machine. This
provides a single place to check whether a requested target state is
allowed to be requested by checking if the value is "in" the keys of
this dictionary. It also provides a mapping for two phrases which are
currently inconsistent:
  target:deleted -> action:delete
  target:active  -> action:deploy

It updates the API controller for nodes.py without changing the HTTP
status codes which are returned in exception cases. The changes in unit
tests now prepare the node appropriately, but do not change the
assertions being tested.

Juno allowed the rebuild of a failed deploy to be initiated by PUTting
a target state of "active"; this behavior, while inconsistent with the
new state machine, is added to the state model for compatibility, now
that stronger checking is done within the API service, based on the
actual modelling of state transitions.

Implements: blueprint new-ironic-state-machine

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>
Change-Id: Id0af5a607b7432d287eef643c1e30e324dc4f879