Adds support for deploying whole disk images
Ironic decides whether it should deploy whole disk images or partition
based images based on the presence of a kernel and a ramdisk.

Partition based images are still deployed the way it is currently. No change
made in the deployment workflow of partition based images.

Whole Disk Images are dumped directly on the lun of the disk presented so that
they can be localbooted.

Separate PXE-Config entries have been added for whole disk images which when
used to deploy whole disk images chainloads onto the hard drive that was exposed
by the deploy ramdisk.

Partially-Implements blueprint whole-disk-image-support

Change-Id: I4478eff302a0ffeb3e9f2077a41170671863c478