Conductor errors if enabled_drivers are not found
The conductor service should fail during startup if a configured driver,
eg. one that is specified in "enabled_drivers", is not found. At the
moment, it merely ignores missing drivers.

In addition to adding a check that all configured drivers are loaded,
this patch re-arranges some of the initialization of ConductorManager
such that an exception during init_host() does not result in
secondary exceptions being logged and obfuscating the cause, namely,
that no drivers were loaded.

Closes-bug: #1417954

Change-Id: I6a96f4b49891c92859479656fe36d2bbd38fe602