Implement Cleaning States
Adds cleaning states to the state machine and implements _do_node_clean
in the conductor. Changes DELETING->AVAILABLE and MANAGEABLE->AVAILABLE
transitions to always go through the cleaning process. Cleaning can
be disabled with a config option added in a later patch to maintain
the same functionality as in Juno had with ACTIVE->DELETING->NOSTATE.

Implements blueprint implement-cleaning-states
Change-Id: Ie968e5b3d45f417e98da6b3d985de5bfd851c364