PXE driver: Deprecate pxe_deploy_{ramdisk, kernel}
Since we now can use the IPA ramdisk as the deploy ramdisk for both
the PXE and Agent drivers, it makes sense to support passing the same
parameters to the node.driver_info on both drivers type. This patch is
deprecating the "pxe_deploy_ramdisk" and "pxe_deploy_kernel" parameters
for the PXE drivers in favor of the "deploy_ramdisk" and "deploy_kernel"
parameters that is already used by the Agent drivers.

We are logging a warning when the "pxe_" namespaced parameters are used
and also saying it's deprecated when get_options() is invoked via the API.

DocImpact
Change-Id: Ib2c4eba57bd5535310bf5b026a7ab01932ec4300