Make try block shorter for _make_password_file
The yield statement should not be inside the try block since it will be
used as a 'with' statement.  When an exception is raised inside the with
block it gets caught and was raising a PasswordFileFailedToCreate
exception.

It is not required to set the mkstemp() file as R/W only by the user as
this is what mkstemp() does.

Change-Id: Id4b0aaabd6bc493874232c6445340787b7ca3ae6