Rename _continue_deploy() to pass_deploy_info()
For historic reasons the method called by the bash ramdisk to continue the
deployment was called _continue_deploy() (in code) but is exposed on the
API as "pass_deploy_info" (.../vendor_passthru/pass_deploy_info). The
vendor interface for IPA contains a method called continue_deploy()
(without the '_' prefix) and since we got the pxe_* drivers to work
with both ramdisks the name of both methods was causing some confusion,
because the only difference between them was the '_' prefix but they
are not related at all.

So this patch is now renaming the _continue_deploy() methods to
pass_deploy_info (same name it's exposed on the API) to improve the code
readibility of the vendor interface.

Change-Id: I4d909a31585eedbec2df2cd3575f658c1cb0d0aa