Enable cleaning by default
This patch enables cleaning by default.

It is always safer to erase data after an instance has been deleted, and
as such, this is a sane production default.

The only potential impact of this option being enabled
is that deletion may (depending on hardware) take longer.

This commit also updates the help string to describe the only known
production situation where disabling this option makes sense -- that is,
where there is only a single, trusted tenant.

Implements blueprint implement-cleaning-steps

Depends-On: I77081165a80491da3e66d8a4554b6d71fc3d9353
Change-Id: I4ed4fbc8c4d88988add953307f8871cbfa33971f