is_whole_disk_image might not exist for previous instances
For newly created instances, we assume that 'is_whole_disk_image'
should exist as it is populated by the conductor before every
deploy.  But it might not exist for instances create before
whole disk image support (as it was introduced in this feature).
We should assume that it's a partition image during tear down or
takeover of these nodes and not fail.

Closes-bug: 1432093
Change-Id: If347a6d64be8bf4959195b9c68293c5ee9d40b48