Hide new fields via single method
With the introduction of microversions, node fields that are added in
a particular (micro)version should not be made visible to requests
for older versions. This patch puts the logic for handling this for
all such new fields in one method: hide_fields_in_newer_versions().

This should make it easier for developers to know where to add code,
and will (hopefully) prevent multiple methods from popping up for
each new set of fields.

This also removes the exception handling in allow_logical_names()
because it is not needed, since it is no longer being called in a
method where a pecan.request might not be available (eg, when
generating a sample of a Node for the documentation).

Change-Id: Id6aeb3543d447113a8cf702f2e85d8a9ca3858f3