Allow node.instance_uuid to be removed during cleaning
As soon as Nova moves the node's provision state to DELETING, it
attempts to remove the node's instance UUID. Ironic throws an error
because it considers the node in a state transition. If cleaning
takes longer than the Nova timeout, the delete will error out.

The fix is to allow node.instance_uuid removal while the node is in
CLEANING state. In L, we should fix the Nova driver to not remove
node.instance_uuid, and clear it after tear down when we clear out
node.instance_info.

Also allows updates to nodes in CLEANFAIL to match the other states
in UPDATE_ALLOWED_STATES, and if the node fails CLEANING quickly, the
node.instance_uuid removal could happen here.

Change-Id: I46771041f90a1d52f6b2d8e107ca10b650b720c1
Closes-Bug: 1435605