Attach ilo_boot_iso only if node is active
Currently iLO drivers attach instance_info/ilo_boot_iso
(if it exists) to the node everytime it powers on the node.
This commit changes the ilo_boot_iso to be attached
only if the node is in active state.

This is for scenarios like instance rebuild where
instance_info/ilo_boot_iso might exist, but it shouldn't
be attached while powering on the node (during rebuild,
the node should boot from deploy_iso again to deploy the
new image).

Closes-bug: 1435959
Change-Id: I0fb0b0c016e52f88807d13c45c03127b7efdf4fb