DRAC: power on during reboot if powered off
The DRAC driver would throw an exception if a reboot was
requested when the node was powered off. The node is now
powered on in that case, instead of raising an exception.

This is as per the behaviour described in 
Ia527bc6c7a9b04f38786069cf676265bae37a44b.

Closes-bug: #1438166
Change-Id: I4ec95b79ae4c8c0af0fc01a12a59e97643e4b2c9