Remove inspect_ports from ilo inspection
This commit removes the driver_info parameter
'inspect_ports' which was required for ilo driver
inspection. After this change, all the NICs detected
by iLO will be enrolled as Ironic ports and operators
are supposed to delete the Ironic ports for unused
NICs  manually (deploy may fail if all Ironic ports
are enrolled due to bug 1405131).

Closes-bug: 1438546
Change-Id: I37eb1dda649d4d21681555339c9ece8a18216f95