Disable XML now that we have WSME/Pecan support
We've had a desire to disable XML from the Ironic API for some
time, however to do so needed support in WSME/Pecan.  That's now
been added, so we can implement the Ironic side.

To do so, we wrap wsmeext.pecan.expose() and replace all the decorators
with a call to ironic.api.expose.expose().

Change-Id: I94d0387722225c4356c867f63b6f1a61356f317d
Closes-bug: 1271317