Clarify message for power action during cleaning
If someone tries to perform a power action on a node that is being
cleaned, they will get a message like 'The requested action
"power off" can not be performed on node "<identifier>" while
it is in state "power on".'

That message doesn't explain the problem. This changes the message
to '... while it is in state "cleaning".'

This addresses a nit in commit
I37dfe962e6c28ed9714f4d3ba8dc55894fc162d4.

Implements blueprint implement-cleaning-states

Change-Id: I5983ce598ec1e82fcc5579b2ffaea27e68e259aa