Add whole disk image support for iscsi_irmc driver
This adds local boot capability to whole disk deployed image
in addition to partition deployed image in iscsi_ilo driver.
The primary change is to set a bare metal node to boot from
local disk at the end of deployment if the deployed image is a
whole disk image.

Implements: blueprint whole-disk-image-support
Change-Id: Ia7d0156511048eb45b52d573d25a546d84cc6875