Improve root partition size check in deploy_partition_image
There is a check in deploy_partition_image which ensures that size of
the image being deployed is greater than or equal to root partition
size. It is being checked on driver.deploy step by calling
iscsi_deploy.check_image_size by in-tree drivers, but if some
out-of-tree driver would use this function without calling
iscsi_deploy.check_image_size beforehand the logic would be
incorrect, as deploy_partition_image will silently change root_mb
value and proceed.

In this change it is proposed to throw an exception instead in this
case.

Closes-bug: #1454263

Change-Id: Iecd59d938762089ab639066fa994492a64e92937