Return HTTP 400 for invalid sort_key
HTTP 500 (Internal Server Error) was being returned when a
request (eg node-list) specified an invalid sort_key. This is
because a DB exception was being raised that we weren't
handling.

We now raise an InvalidParameterValue exception with a message
that indicates that the sort key is invalid. The response is now
an HTTP 400 (Bad Request).

Change-Id: I237856ccb10e395bf2b42307804f607963b5dcbb
Closes-Bug: #1450552