Add section header to state machines page
This adds a section header so that a "Table of Contents" is
displayed on the rendered page. This makes the page consistent
with the rest of our developer documentation, since the others
all have a Table Of Contents.

I also added the word 'provisioning' to qualify what kind of
states were being depicted here. (As opposed to power states.)

Change-Id: If71679d9a5ee795c6c3a9d912f371ea04cea2753