API: Get a subset of fields from Nodes
Sometimes we don't care about all fields in a node (or a collection
of them) and only need a particular subset of fields for a particular
operation. This can save a lot of bandwidth depending on the information
contained in the nodes, for example, a node which contains a config drive
bsae64 encoded in it's instance_info field can be pretty heavy (a
collection of them ...)

This patch allows users to specify a list of fields that will be returned
when GET'ing a node or a collection of nodes.

E.g

Get the uuid, provision_state and power_state of a particular node:

  GET v1/nodes/<uuid>?fields=uuid,provision_state,power_state

Get the uuid, provision_state and power_state of a collection of nodes:

  GET v1/nodes?fields=uuid,provision_state,power_state

Get the uuid, provision_state and power_state of a particular node associated
with a particular instance:

  GET v1/nodes?fields=uuid,provision_state,power_state&instance_uuid=<instance uuid>

Since we are introducing a new feature in the API this patch bumps the
micro version to 1.8

APIImpact
Partial-Bug: #1466495
Related-Bug: #1458934
Change-Id: Iee3b8821afeb670848d4030a5d1a29b356bb5559