Prevent node delete based on provision, not power, state
This patch removes a conditional that was preventing a node from being
deleted (from the DB) based on its power state. We should not care about
what power state the node currently is before deleting it but instead we
should care about deleting a node depending on its provision state. Say
a node is active or cleaning or zapping, etc... We don't want to delete
the node on those states because there are some actions being made at
those states.

Closes-Bug: #1461115
Change-Id: I9778898477c8129e1e98be066695bd96f8f660b8