Don't raise exception from set_failed_state()
deploy_utils.set_failed_state() shouldn't raise any exceptions because
it is used to handle a failed deployment, and we want the node's
last_error to be set.

Instead, in the very rare (if ever) case that the call to
task.process_event() fails, we'll log an exception and continue.

Change-Id: I8f02577de5a5bce34608882f61b5b26f5b998fad
Closes-Bug: #1461670