Add vendor-passthru to attach and boot an ISO
This commit adds a new vendor-passthru 'boot_into_iso' 
in iscsi_ilo driver to attach and boot from an ISO
located in Glance or an HTTP(s) URL.  This vendor passthru
can be invoked only when the node is in manage state or
in maintenance mode.  This vendor-passthru can help in
debugging issues with bare metal node.

Partial-Bug: 1464168
Change-Id: I75c5363b1df13b12c8f025a4b45de02367229f8f