get_supported_boot_devices() returns static device list
Returning a static device list currently is only applicable
to x86 based architectures. If a node references for instance
sparc architecture, then "pxe" boot is not supported, however
"wanboot" is. The driver referenced by the node should be able
to return a custom boot device list dependent on the architecture
of the node. However to achieve this the task needs to be passed
as an argument in get_supported_boot_devices()

Change-Id: I5ae70ee84309dffb7d7bca4ac5cba28b01b7c414
Closes-Bug: #1391598
Implements: blueprint supported-boot-device-list