Don't support deprecated drivers' vendor_passthru
In Kilo, a common mechanism was added that all drivers use for exposing
their vendor methods. Drivers' own implementations of vendor_passthru()
method was deprecated, and backwards compatibility was maintained for any
out-of-tree drivers that still implemented their own version of the
vendor_passthru() method.

This change removes support for that backwards compatibility.

Change-Id: I5e134e42449f2bbeefee1f9cb1712f59fa7a2b9f