Fix failing unit tests under py34
When running the unit tests under Python >= 3.4.2, several of them
fail with "TypeError: 'xxx' object is not iterable".  This is due to a
bug in mock: http://bugs.python.org/issue23661.

The workaround to this bug is to make side effects that are exceptions
iterable.  This change makes the unit tests work on both py27 and py34.

Change-Id: I4bcda02bb5756fea2a8daecc238ef7439ad43700
Partial-Bug: #1463867