Clean nodes stuck in DEPLOYING state when ir-cond restarts
When a conductor managing a node dies abruptly mid deployment (due some
OMM killer, power outage, etc...) the node will get stuck in the
DEPLOYING state. There is no timeout for nodes in DEPLOYING state, and
if a conductor with the same hostname comes back online, it merely
releases the locks held by that hostname. This is not enough, however,
because the API service prevents changes to nodes in DEPLOYING state. We
need the conductor to update the node's provision state from DEPLOYING
to DEPLOYFAIL. The only work around is directly modifying the DB today.

This patch mitigates this problem by adding a step during conductor
startup: after releasing locks previously held by this conductor, it
will now query the database to identify nodes in the DEPLOYING state
that no longer have a valid lock. These nodes are then transitioned to
the DEPLOYFAIL state.

Nodes locked by other conductors, or in the DEPLOYING state with a valid
lock, will remain untouched.

Partial-Bug: #1461937
Change-Id: I59638661e35b80ef6c07a204889a45882acf5684