Update cached images based on update time
Currently, if non-glance image is used in ironic, and contents of this
image are updated on HTTP(S) server or on file system available
locally (except for the case when it is the file system where image
cache is located), ironic won't be able to determine that the image
was updated and will not update the cache.

In this change it is proposed to use last modification date to
determine if an image is up to date. For file system it can be
accomplished with standard operating system means, and for HTTP(S)
images 'Last-Modified' HTTP header can be used.

Closes-bug: #1459328
Change-Id: Ia3e6b6072fdf7a4bfffdc3fbaf31103605c7ff5b