Save disk layout information when deploying
When using 'preserve ephemeral', we save the disk layout information
so that if a rebuild is requested, we can check that the disk layout
hasn't been changed by the user.

This information (ephemeral_gb, swap_mb, and root_gb) was being
saved in the node's driver_internal_info['instance'] by
iscsi_deploy.parse_instance_info(), which may be called several
times and by e.g. validate(). However, the information should only
be saved when a node is being deployed to, and not before.

Now the information is only saved when a node is being deployed to.

This patch also includes some updates to docstrings and clean up of
code that was introduced in commit
Iecdb8db37b22c37275be3197950bbf899aa62a53.

Change-Id: If8e6a1fefe0f48e5301364143ce86e58e8740635
Closes-Bug: #1327260