Do not use "private" attribute in AuthTokenMiddleware
There is a convention that is followed Python code, a name prefixed
with an underscore should be treated as a non-public part of the API
It should be considered an implementation detail and subject to
change without notice. This patch removes usage of "_app" attribute
of base class from AuthTokenMiddleware.

Change-Id: I8d7fa42796ccfd0173335c1d515c789bf2be965b