Periodically checks the status of nodes in DEPLOYING state
Periodically checks the nodes in DEPLOYING and the state of the conductor
deploying them. If we find out that a conductor that was provisioning
the node has died we then break release the node and gracefully mark
the deployment as failed.

A new method called "get_offline_conductors" was added to the database
api which return a list of conductor hostnames that are offline.

Closes-Bug: #1461937
Change-Id: I293d3ff18e949b305e6404b8d3ceb75e29b88356